+++ relay-filter/src/config.rs
@@ -136,6 +136,20 @@ pub struct ErrorMessagesFilterConfig {
     pub patterns: GlobPatterns,
 }
 
+/// Configuration for transaction name filter.
+#[derive(Clone, Debug, Default, Serialize, Deserialize)]
+pub struct IgnoreTransactionsFilterConfig {
+    /// List of patterns for ignored transactions that should be filtered.
+    pub patterns: GlobPatterns,
+}
+
+impl IgnoreTransactionsFilterConfig {
+    /// Returns true if no configuration for this filter is given.
+    pub fn is_empty(&self) -> bool {
+        self.patterns.is_empty()
+    }
+}
+
 impl ErrorMessagesFilterConfig {
     /// Returns true if no configuration for this filter is given.
     pub fn is_empty(&self) -> bool {
@@ -210,6 +224,13 @@ pub struct FiltersConfig {
     /// Configuration for the releases filter.
     #[serde(default, skip_serializing_if = "ReleasesFilterConfig::is_empty")]
     pub releases: ReleasesFilterConfig,
+
+    /// Configuration for ignore transactions filter.
+    #[serde(
+        default,
+        skip_serializing_if = "IgnoreTransactionsFilterConfig::is_empty"
+    )]
+    pub ignore_transactions: IgnoreTransactionsFilterConfig,
 }
 
 impl FiltersConfig {
@@ -223,6 +244,7 @@ impl FiltersConfig {
             && self.legacy_browsers.is_empty()
             && self.localhost.is_empty()
             && self.releases.is_empty()
+            && self.ignore_transactions.is_empty()
     }
 }
 
@@ -260,6 +282,9 @@ mod tests {
             releases: ReleasesFilterConfig {
                 releases: [],
             },
+            ignore_transactions: IgnoreTransactionsFilterConfig {
+                patterns: [],
+            },
         }
         "###);
         Ok(())
@@ -292,6 +317,9 @@ mod tests {
             localhost: FilterConfig { is_enabled: true },
             releases: ReleasesFilterConfig {
                 releases: GlobPatterns::new(vec!["1.2.3".to_string()]),
+            },
+            ignore_transactions: IgnoreTransactionsFilterConfig {
+                patterns: GlobPatterns::new(vec!["*health*".to_string()]),
             },
         };
 
@@ -330,6 +358,11 @@ mod tests {
           "releases": {
             "releases": [
               "1.2.3"
+            ]
+          },
+          "ignoreTransactions": {
+            "patterns": [
+              "*health*"
             ]
           }
         }